+++ src/core_modules/capture-core/components/WidgetsChangelog/common/Changelog/Changelog.container.js
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react';
 import { Modal } from '@dhis2/ui';
-import { useChangelogData } from '../hooks';
+import { useChangelogData, useClientDataValues } from '../hooks';
 import { ChangelogComponent } from './Changelog.component';
 import { CHANGELOG_ENTITY_TYPES } from './index';
 import { LoadingMaskElementCenter } from '../../../LoadingMasks';
@@ -14,7 +14,7 @@ type Props = {
     close: () => void,
     dataItemDefinitions: ItemDefinitions,
     programId?: string,
-}
+};
 
 export const Changelog = ({
     entityId,
@@ -25,9 +25,11 @@ export const Changelog = ({
     dataItemDefinitions,
 }: Props) => {
     const {
-        records,
+        rawRecords,
         pager,
-        isLoading,
+        isLoading: isChangelogLoading,
+        page,
+        pageSize,
         setPage,
         setPageSize,
         sortDirection,
@@ -36,10 +38,23 @@ export const Changelog = ({
         entityId,
         entityType,
         programId,
-        dataItemDefinitions,
     });
 
-    if (isLoading) {
+    const {
+        processedRecords,
+        isLoading: isProcessingLoading,
+    } = useClientDataValues({
+        rawRecords,
+        dataItemDefinitions,
+        entityId,
+        entityType,
+        programId,
+        sortDirection,
+        page,
+        pageSize,
+    });
+
+    if (isChangelogLoading || isProcessingLoading) {
         return (
             <Modal onClose={close}>
                 <LoadingMaskElementCenter />
@@ -51,7 +66,7 @@ export const Changelog = ({
         <ChangelogComponent
             isOpen={isOpen}
             close={close}
-            records={records}
+            records={processedRecords}
             pager={pager}
             setPage={setPage}
             setPageSize={setPageSize}
