+++ src/core_modules/capture-core/components/WidgetsChangelog/common/hooks/index.js
@@ -1,3 +1,4 @@
 // @flow
 
 export { useChangelogData } from './useChangelogData';
+export { useClientDataValues } from './useClientDataValues';
